Add files via upload
Fruitful academic collaborations have become increasingly more important for solving scientific problems, participating in research projects, and im-proving productivity. As such, frameworks for recommending suitable col-laborators are attracting extensive attention from scholars. In an effort to improve on the current solutions, we developed an approach that produces recommendations with better precision, recall, and accuracy. Our strategy is to leverage the benefits of the two most common similarity indicators for collaborator recommendation – research interests and co-authorship net-work topology into one unified framework. A Word2Vec model creates word embeddings of research interests, which solves the problem of calcu-lating similarity solely based on co-occurrence, not context, while a Node2Vec model automatically extracts and learns the topological features of a co-authorship network, moving beyond just local features to capture global network features as well. The two similarity measures are then fused with CombMNZ resulting in a ranked list of recommended collaborators for the target scholar(s). The workings of the framework and its benefits are demonstrated through a case study on academics in the field of intelligent driving and a comparison with the two most commonly used baselines: Random Walk with Restart (RWR) and Latent Dirichlet Allocation (LDA). Our framework should be of benefit for academics, research centers, and private-enterprise R&D managers to find partners, so as to achieve the ulti-mate goal of completing research projects, solving scientific problems, and promoting discipline development and progress.